requirements.txt
anyio==3.6.2
bcrypt==4.0.1
click==8.1.3
colorama==0.4.6
distlib==0.3.6
dnspython==2.3.0
email-validator==2.0.0.post2
Faker==18.4.0
fastapi==0.94.1
filelock==3.9.1
gunicorn==20.1.0
h11==0.14.0
idna==3.4
motor==3.1.2
platformdirs==3.1.1
pydantic==1.10.6
PyJWT==2.6.0
pymongo==4.3.3
python-dateutil==2.8.2
python-dotenv==1.0.0
six==1.16.0
sniffio==1.3.0
starlette==0.26.1
typing_extensions==4.5.0
uvicorn==0.21.0
virtualenv==20.21.0
